Revert "Adding session and upload lengths to request headers of .mparticle requests."
This reverts commit 3baee6fa1f76dee0a54aed19b6f655f519684868.

+++ sdk/src/main/java/com/mparticle/ConfigManager.java
@@ -129,7 +129,7 @@ class ConfigManager {
         return localPrefs.mSecret;
     }
 
-    public int getUploadInterval() {
+    public long getUploadInterval() {
         if (localPrefs.debug) {
             return 1000 * localPrefs.debugUploadInterval;
         } else {
